checked input

+++ LIBFT/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: akharkho <akharkho@student.42.fr>          +#+  +:+       +#+        */
+/*   By: kali <kali@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/27 17:56:41 by akharkho          #+#    #+#             */
-/*   Updated: 2025/01/28 18:07:28 by akharkho         ###   ########.fr       */
+/*   Updated: 2025/01/29 11:17:45 by kali             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_stack
 t_stack	*ft_newnode(int value, int index);
 void	ft_nodeadd_back(t_stack **lst, t_stack *new);
 t_stack	*ft_newnode(int value, int index);
-int		ft_atoi(const char *str);
+long		ft_atoi(const char *str);
 int		ft_isalpha(int c);
 int		ft_isdigit(int c);
 int		ft_strlen(char *str);
